pyproject.toml
[project]
name = "roary"
version = "0.1.0"
description = "Autonomous multi-agent GitHub content repurposer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.1",
    "crewai>=1.12.2",
    "fastapi>=0.135.3",
    "gitpython>=3.1.44",
    "langchain-chroma>=1.0.0",
    "langchain-huggingface>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "pydantic>=2.0",
    "pydantic-ai>=1.31.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=3.0.0",
    "uvicorn>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/roary"]
